Allow RSS Feed Title to be customized
Thanks for the excellent conversion from a starter to a theme! If it's alright I'd love to be able to adjust the title for the RSS feed, but it seemed excessive to overwrite the entire plugin just to do that. It was also tempting to use the `siteMetadata.title`, but I personally wanted to have a different title for RSS feeds since it's just the "Blog" that gets added there.